# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CourseHub API",
        "description": "Complete API documentation for CourseHub platform",
        "contact": {
            "name": "CourseHub Support",
            "email": "admin@coursehub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/courses/{course}/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Retrieve a list of course lessons",
                "description": "Retrieve a paginated list of course lessons.",
                "operationId": "235f3dcd29c974b5dfacc6f125487fc1",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "math-101"
                        }
                    },
                    {
                        "name": "filter[search]",
                        "in": "query",
                        "description": "Search lessons by title.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort lessons by field. Use \"-\" prefix for descending order.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "-title",
                                "position",
                                "-position",
                                "created_at",
                                "-created_at"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson list retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LessonResponse"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User does not have permissions."
                    },
                    "404": {
                        "description": "Course not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create lessons",
                "description": "Create a new lesson for a specific course.",
                "operationId": "3688868ff5b41c7b976a7a0c10e012d3",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLessonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/courses/{course}/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get lesson",
                "description": "Retrieve a specific lesson for a specific course.",
                "operationId": "44edac41fc34e68e26d87d0e6a01ba10",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Course or Lesson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    },
                    {}
                ]
            },
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update lesson",
                "description": "Update a specific lesson for a specific course.",
                "operationId": "19eef4647ae27f9c616744c34e7b7ccb",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLessonRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Course or Lesson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete lesson",
                "description": "Delete a specific lesson from a specific course.",
                "operationId": "f0874920178e45bebf1598a0907b5698",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson identifier (slug)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lesson deleted"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Course or Lesson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LessonResponse": {
                "title": "Data of a specific lesson.",
                "required": [
                    "id",
                    "course_id",
                    "title",
                    "slug",
                    "position",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier of the lesson.",
                        "type": "integer",
                        "example": 1
                    },
                    "course_id": {
                        "description": "Course unique identifier of the lesson.",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Title of the lesson.",
                        "type": "string",
                        "example": "Introduction to Algebra"
                    },
                    "slug": {
                        "description": "Slug of the lesson.",
                        "type": "string",
                        "example": "introduction-to-algebra"
                    },
                    "position": {
                        "description": "Position of the lesson.",
                        "type": "integer",
                        "example": 1
                    },
                    "content": {
                        "description": "Content of the lesson.",
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/OfflineLessonResponse"
                            },
                            {
                                "$ref": "#/components/schemas/OnlineLessonResponse"
                            },
                            {
                                "$ref": "#/components/schemas/VideoLessonResponse"
                            }
                        ]
                    },
                    "created_at": {
                        "description": "Creation date of the lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Modification date of the lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-10T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "OfflineLessonResponse": {
                "title": "Offline Lesson Response",
                "description": "Content of a specific offline lesson.",
                "required": [
                    "start_time",
                    "end_time",
                    "address",
                    "room_number"
                ],
                "properties": {
                    "start_time": {
                        "description": "Start time of the offline lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2027-02-01T10:00:00Z",
                        "nullable": true
                    },
                    "end_time": {
                        "description": "End time of the offline lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2027-02-01T12:00:00Z",
                        "nullable": true
                    },
                    "address": {
                        "description": "Address of the offline lesson.",
                        "type": "string",
                        "example": "123 Main St, Springfield",
                        "nullable": true
                    },
                    "room_number": {
                        "description": "Room number of the offline lesson.",
                        "type": "string",
                        "example": "Room 101",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OnlineLessonResponse": {
                "title": "Online Lesson Response",
                "description": "Content of a specific online lesson.",
                "required": [
                    "start_time",
                    "end_time",
                    "meeting_link"
                ],
                "properties": {
                    "start_time": {
                        "description": "Start time of the online lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2027-02-01T10:00:00Z",
                        "nullable": true
                    },
                    "end_time": {
                        "description": "End time of the online lesson.",
                        "type": "string",
                        "format": "date-time",
                        "example": "2027-02-01T12:00:00Z",
                        "nullable": true
                    },
                    "meeting_link": {
                        "description": "Link of the online lesson.",
                        "type": "string",
                        "format": "uri",
                        "example": "https://meet.example.com/lesson123",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "VideoLessonResponse": {
                "title": "Video Lesson Response",
                "description": "Content of a specific video lesson.",
                "required": [
                    "video_url",
                    "provider"
                ],
                "properties": {
                    "video_url": {
                        "description": "Link of the video lesson.",
                        "type": "string",
                        "format": "uri",
                        "example": "https://videos.example.com/lesson123.mp4",
                        "nullable": true
                    },
                    "provider": {
                        "description": "Provider of the video lesson.",
                        "type": "string",
                        "example": "youtube",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StoreLessonRequest": {
                "title": "Store Lesson Request",
                "description": "Schema for creating a lesson depending on course type",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/StoreOfflineLessonRequest"
                    },
                    {
                        "$ref": "#/components/schemas/StoreOnlineLessonRequest"
                    },
                    {
                        "$ref": "#/components/schemas/StoreVideoLessonRequest"
                    }
                ]
            },
            "StoreOfflineLessonRequest": {
                "title": "Store Offline Lesson Request",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreLessonBaseRequest"
                    },
                    {
                        "required": [
                            "start_time",
                            "end_time",
                            "address",
                            "room_number"
                        ],
                        "properties": {
                            "start_time": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2027-02-01T10:00:00Z"
                            },
                            "end_time": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2027-02-01T12:00:00Z"
                            },
                            "address": {
                                "type": "string",
                                "maxLength": 255,
                                "example": "123 Main St, Springfield"
                            },
                            "room_number": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "Room 101"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "StoreOnlineLessonRequest": {
                "title": "Store Online Lesson Request",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreLessonBaseRequest"
                    },
                    {
                        "required": [
                            "start_time",
                            "end_time",
                            "meeting_link"
                        ],
                        "properties": {
                            "start_time": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2027-02-01T10:00:00Z"
                            },
                            "end_time": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2027-02-01T12:00:00Z"
                            },
                            "meeting_link": {
                                "type": "string",
                                "format": "uri",
                                "maxLength": 2048,
                                "example": "https://meet.example.com/lesson123"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "StoreVideoLessonRequest": {
                "title": "Store Video Lesson Request",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/StoreLessonBaseRequest"
                    },
                    {
                        "required": [
                            "video_url",
                            "provider"
                        ],
                        "properties": {
                            "video_url": {
                                "type": "string",
                                "format": "uri",
                                "maxLength": 2048,
                                "example": "https://videos.example.com/lesson123.mp4"
                            },
                            "provider": {
                                "type": "string",
                                "maxLength": 50,
                                "example": "youtube"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "UpdateLessonRequest": {
                "title": "Update Lesson Request",
                "description": "Schema for updating a lesson depending on course type",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpdateOfflineLessonRequest"
                    },
                    {
                        "$ref": "#/components/schemas/UpdateOnlineLessonRequest"
                    },
                    {
                        "$ref": "#/components/schemas/UpdateVideoLessonRequest"
                    }
                ]
            },
...
            "Lesson": {
                "title": "Lesson schema",
                "description": "Details of a lesson returned by the API",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/LessonData"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}